[#153] Move PackageVersion to elementary; code simplification
