+++ iceoryx2-bb/elementary/src/lib.rs
@@ -19,6 +19,7 @@ pub mod bump_allocator;
 pub mod lazy_singleton;
 pub mod math;
 pub mod owning_pointer;
+pub mod package_version;
 pub mod pointer_trait;
 pub mod relocatable_container;
 pub mod relocatable_ptr;
